Update README.md
Just write this for my self to use in scriptmerge plugin for joomla I could not find it on the internet so I thought lets help my fellow website builders or plugin writers. It's verry simple code but so am I. I do not now how Github works and to lazy to read so there is a big change this post is totaly wrong the code is working do. I think if you experince it not working let my now and if this post is placed ad the wrong website tell me..