Create Remove comments from javascript in PHP RegEx
These are 115 lines off code that work together also works with mootools. Tested it on 800.000 lines of code and comments. works fine. 16-01-2016..! This is the code without the comments in it.!!!!  This Handles all Javascript also HTML text in javascript.  Works on already compressed files as on regex code in javascript.